refactor(GeneratorCommand): remove build directory argument
Specifying a different build directory will only change the location of the index.html file.
All other assets in the "public/" directory would have to be copied to the differing build directory too.
This may be a feature for the future.
Therefore, the argument is removed for the time being.